% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory_split.R
\name{directory_split}
\alias{directory_split}
\title{Wrapper for pdf split function}
\usage{
directory_split(
  directory,
  keyword,
  ignore_case = FALSE,
  keyword_combination = "OR",
  combine_pages = TRUE,
  dir_output = NULL
)
}
\arguments{
\item{directory}{The directory to all pdf files to apply the pdf_split function.}

\item{keyword}{The keyword(s) to be used foe extracting pages. Multiple
keywords can be specified with a character vector.}

\item{ignore_case}{TRUE/FALSE Indicates whether keyword is case sensitive.
Default is FALSE meaning that the case of the keyword is
matched exactly.}

\item{keyword_combination}{"OR"/"AND" Determines how multiple keywords in
keyword vector are matched. "OR" means match any
keyword in the keyword vector. "AND" means match
all keywords must be matched.}

\item{combine_pages}{TRUE/FALSE Indicates if you all matched pages should be
output as a single pdf or as individual pdf document.
Default is TRUE.}

\item{dir_output}{Specify path to the directory to store output file(s).
Default is NULL and files will be saved to
output directory created within input directory.}
}
\value{
Path to pdf file containing extracted page(s) when combine_pages = TRUE.
Otherwise, TRUE when extractions are successful.
}
\description{
This will apply the pdf_split function over all PDF files within a
directory.
}
\examples{

directory <- system.file("extdata", package = "pdfsplit")
dir_outs <- file.path(getwd(),"output")
dir.create(dir_outs)
directory_split(directory,keyword = "test", dir_output = dir_outs)
unlink(dir_outs,recursive = TRUE)
}
