% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyword_pages.R
\name{keyword_pages}
\alias{keyword_pages}
\title{Identify pdf pages for keyword}
\usage{
keyword_pages(path, keyword, ignore_case = FALSE, simplify = FALSE)
}
\arguments{
\item{path}{A direct path to a pdf file}

\item{keyword}{The keyword(s) to be used foe extracting pages. Multiple
keywords can be specified with a character vector.}

\item{ignore_case}{TRUE/FALSE Indicates whether keyword is case sensitive.
Default is FALSE meaning that the case of the keyword is
matched exactly.}

\item{simplify}{TRUE/FALSE Indicates whether to return a list for
multiple keywords or return a data frame.Default is FALSE}
}
\value{
A list of keyword and the pdf pages they are located.
Returns a dataframe when simplify is TRUE for multiple keywords.
}
\description{
Leverage pdf_text from pdftools to search the pages a keyword
is located in a pdf file
}
\examples{
# Single keyword
file <- system.file("extdata", "1406.4806.pdf", package = "pdfsplit")
keyword_pages(file, "test", simplify = FALSE)

# Multiple keywords
file <- system.file("extdata", "1406.4806.pdf", package = "pdfsplit")
keyword_pages(file, c("test", "performance"))

## Simplified output
keyword_pages(file, c("test", "performance"), simplify = TRUE)
}
