% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_split.R
\name{pdf_split}
\alias{pdf_split}
\title{Extract pdf pages based on keyword}
\usage{
pdf_split(
  path,
  keyword,
  ignore_case = FALSE,
  keyword_combination = "OR",
  combine_pages = TRUE,
  file_output = NULL
)
}
\arguments{
\item{path}{A direct path to a pdf file}

\item{keyword}{The keyword(s) to be used foe extracting pages. Multiple
keywords can be specified with a character vector.}

\item{ignore_case}{TRUE/FALSE Indicates whether keyword is case sensitive.
Default is FALSE meaning that the case of the keyword is
matched exactly.}

\item{keyword_combination}{"OR"/"AND" Determines how multiple keywords in
keyword vector are matched. "OR" means match any
keyword in the keyword vector. "AND" means match
all keywords must be matched.}

\item{combine_pages}{TRUE/FALSE Indicates if you all matched pages should be
output as a single pdf or as individual pdf document.
Default is TRUE.}

\item{file_output}{Specify path of the output file(s).
Default is NULL and files will be saved to input path.}
}
\value{
Path to pdf file containing extracted page(s) when combine_pages = TRUE.
Otherwise, TRUE when extractions are successful.
}
\description{
Uses pdf splitting capability of the pdftools package to extract pages from
a pdf file based on user specified keyword(s).
}
\examples{
# Single keyword
file <- system.file("extdata", "1406.4806.pdf", package = "pdfsplit")
pdf_file <- file.path(getwd(), "output.pdf")
pdf_split(file,keyword = "test", file_output = pdf_file)
unlink(pdf_file)

# Multiple keywords
file <- system.file("extdata", "1406.4806.pdf", package = "pdfsplit")
pdf_file <- file.path(getwd(), "output.pdf")
pdf_split(file,keyword = c("test","performance"),keyword_combination = "OR",
 file_output = pdf_file)
unlink(pdf_file)


}
